Remove generated testing for now
This is a smoke test for deployment which will come in handy later for
verifying the HELM chart is working as intended. For an initial release,
it's more trouble than it's worth. Plus, we need to build in better
testing with our containers so we know they are working correctly.